--- packages/rating-scales/.babelrc.json
@@ -1,17 +1,0 @@
-{
-  "sourceType": "unambiguous",
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": {
-          "chrome": 100,
-          "safari": 15,
-          "firefox": 91
-        }
-      }
-    ],
-    "@babel/preset-react"
-  ],
-  "plugins": []
-}
