+++ packages/rating-scales/.storybook/preview.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { ChakraProvider } from "@chakra-ui/react";
+
+const preview = {
+  decorators: [
+    (Story) => (
+      <ChakraProvider resetCSS>
+        <Story />
+      </ChakraProvider>
+    ),
+  ],
+};
+
+export default preview;
