+++ packages/rating-scales/package.json
@@ -61,7 +61,7 @@
     "@storybook/addon-onboarding": "^1.0.8",
     "@storybook/blocks": "^7.5.3",
     "@storybook/react": "^7.5.3",
-    "@storybook/react-webpack5": "^7.5.3",
+    "@storybook/react-vite": "^7.5.3",
     "@storybook/testing-library": "^0.2.2",
     "babel-loader": "^8.2.2",
     "core-js": "^3.12.1",
@@ -72,8 +72,7 @@
     "rollup": "^2.48.0",
     "rollup-plugin-analyzer": "^4.0.0",
     "rollup-plugin-terser": "^7.0.2",
-    "storybook": "^7.5.3",
-    "webpack": "^5.37.1"
+    "storybook": "^7.5.3"
   },
   "peerDependencies": {
     "@emotion/react": "~11",
