+++ packages/rating-scales/src/core/Question.stories.tsx
@@ -1,11 +1,8 @@
 import { Question } from "./Question";
 
 export default {
-  title: "Example/Question",
+  title: "core/Question",
   component: Question,
-  parameters: {
-    layout: "centered",
-  },
   tags: ["autodocs"],
   argTypes: {
     textColor: { control: "color" },
@@ -27,31 +24,5 @@ export const Default = {
     fontSize: "1.6em",
     xAlign: "left",
     children: "Your question here",
-  },
-};
-
-export const CenterAligned = {
-  args: {
-    ...Default.args,
-    xAlign: "center",
-  },
-};
-
-export const RightAligned = {
-  args: {
-    ...Default.args,
-    xAlign: "right",
-  },
-};
-
-export const CustomStyle = {
-  args: {
-    textColor: "blue",
-    topMargin: "20%",
-    xMargin: "10%",
-    fontFamily: "Times New Roman",
-    fontSize: "2em",
-    xAlign: "center",
-    children: "Your custom styled question",
   },
 };
