--- packages/rating-scales/src/stories/Page.stories.js
@@ -1,25 +1,0 @@
-import { within, userEvent } from "@storybook/testing-library";
-
-import { Page } from "./Page";
-
-export default {
-  title: "Example/Page",
-  component: Page,
-  parameters: {
-    // More on how to position stories at: https://storybook.js.org/docs/react/configure/story-layout
-    layout: "fullscreen",
-  },
-};
-
-export const LoggedOut = {};
-
-// More on interaction testing: https://storybook.js.org/docs/react/writing-tests/interaction-testing
-export const LoggedIn = {
-  play: async ({ canvasElement }) => {
-    const canvas = within(canvasElement);
-    const loginButton = await canvas.getByRole("button", {
-      name: /Log in/i,
-    });
-    await userEvent.click(loginButton);
-  },
-};
